% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Andrews.R
\docType{data}
\name{Andrews}
\alias{Andrews}
\title{A table of UK bat exit times based on the work of Andrews & Pearson (2022)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 4 columns.
}
\usage{
data(Andrews)
}
\description{
A table of UK bat exit times based on the work of Andrews & Pearson (2022)
the reference is available from:
https://drive.google.com/file/d/1DeGHxyr9-p5XH6R6CRimsmquVD188WY8/view
Relative to sunset the: mean exit time and an exit range based on the standard
deviation is provided.
}
\keyword{datasets}
