% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_colours_default.R
\name{bat_colours_default}
\alias{bat_colours_default}
\title{Assigns UK bat species a default colour in graphs}
\usage{
bat_colours_default(bat_column_to_graph)
}
\arguments{
\item{bat_column_to_graph}{Vector of bat species to be graphed and assigned a species colour}
}
\value{
A named vector of bat species and an assigned default colour
}
\description{
Assigns UK bat species (with scientific names only) a default colour so graphs
can be consistent (i.e. same colour for individual species.)
}
\examples{

bat_colours_default(Lydford$Species)

scales::show_col(bat_colours_default(Lydford$Species))

}
