% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_common_default.R
\name{bat_common_default}
\alias{bat_common_default}
\title{Assigns UK bat common name to a scientific name for default species}
\usage{
bat_common_default(UK_bat_scientific_names)
}
\arguments{
\item{bat_column_to_graph}{Vector of UK bat species with scientific names}
}
\value{
A named vector of bat species with common names
}
\description{
Assigns UK bat common name to a scientific name for default species also
consistent with the default colours for individual species.)
}
\examples{

bat_common_default(Lydford$Species)


}
