% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatExplorer.R
\docType{data}
\name{BatExplorer}
\alias{BatExplorer}
\title{Data exported from Elekon AG Bat Explorer software with call metrics}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 28 rows and 13 columns.
}
\usage{
data(BatExplorer)
}
\description{
Evening transect bat detector survey, sound analysis undertaken with Elekon AG
Bat Explorer software and the data exported as a csv file with the call metrics.
The transect was recorded with a Batlogger M bat detector near the the River
Tavy, West Devon.
}
\keyword{datasets}
