% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date_meta_info.R
\name{date_time_info}
\alias{date_time_info}
\title{Add date and time information}
\usage{
date_time_info(bat_df, time_zone = "Europe/London")
}
\arguments{
\item{bat_df}{bat data frame with DateTime column}
}
\value{
Night, Month, MonthFull, Year, JustTime
}
\description{
Creates useful date and time information (Night, Month, MonthFull, Year, JustTime)
from the DateTime column, as extra columns in the data frame, from the date
and time of the bat observation. The default IANA time zone is "Europe/London".
}
\examples{
bat_data_frame <- tibble(DateTime = Sys.time())
date_time_info(bat_data_frame)

date_time_info(bat_data_frame, time_zone = "Asia/Jakarta")

}
