% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_bat_names.R
\name{duplicate_bat_names}
\alias{duplicate_bat_names}
\title{Check for duplicate names in bat list vector}
\usage{
duplicate_bat_names(look_up_list)
}
\arguments{
\item{look_up_list}{vector of bat names to check for duplicates}
}
\value{
bat character vector of duplicate names
}
\description{
Check for duplicate names in bat list vector
}
\examples{
look_up <- c("45 pip" = "Pipistrellus", "45 pip" = "Pipistrellus", "bleb" = "Plecotus auritus")
duplicate_bat_names(look_up)

}
