% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibernacula_bats.R
\docType{data}
\name{hibernacula_bats}
\alias{hibernacula_bats}
\title{A data.frame of bat activity in a hibernacula.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 8 columns.
}
\usage{
data(hibernacula_bats)
}
\description{
Bat activity in a hibernacula was recorded over a winter period.  This data.frame
details the date and time bat activity was recorded, the species of bat and the length of
time bats were active in the hibernacula. The hibernacula is located in southern Britain.
}
\details{
This data set relates to the hibrenacula_climate data set that records temperature and humidity
for all the buildings observed as a potential hibernacula.
}
\keyword{datasets}
