% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibernacula_climate.R
\docType{data}
\name{hibernacula_climate}
\alias{hibernacula_climate}
\title{A time series data.frame listing temperature, humidity and dew point for potential bat hibernacula.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17312 rows and 6 columns.
}
\usage{
data(hibernacula_climate)
}
\description{
A time series of temperature, humidity and dew point recorded every hour throughout the winter
in potential bat hibernacula located in southern Britain.
}
\keyword{datasets}
