% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kaleidoscope.R
\docType{data}
\name{Kaleidoscope}
\alias{Kaleidoscope}
\title{Data exported from Wildlife Acoustics Kaleidoscope}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 144 rows and 29 columns.
}
\usage{
data(Kaleidoscope)
}
\description{
Anonymised data kindly provided by Richard Graves Associates
}
\keyword{datasets}
