% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lydford.R
\docType{data}
\name{Lydford}
\alias{Lydford}
\title{Lydford Primary School's bat walk observations}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 62 rows and 23 columns.
}
\usage{
data(Lydford)
}
\description{
These are Lydford Primary School's bat walk observations from Wednesday
evening 18^th^ September 2019.  The walk, between the School and the River Lyd,
on the western edge of Dartmoor, started at sunset 6:27pm and continued for
just over an hour.  The observations were made with a Batlogger M bat detector
(https://www.batlogger.com/en/products/batlogger_m/) that records the sound
of the bat as it passes together with the time and postion; using a built-in
Global Positioning System (GPS).  These sounds where then analysed with software
designed to help determine which species made the echolocation;
https://www.batlogger.com/en/products/batexplorer/) was the software used.
}
\keyword{datasets}
