% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_name_lookup.R
\name{make_scientific}
\alias{make_scientific}
\title{Makes bat names scientific}
\usage{
make_scientific(look_up_vector, bat_name_vector)
}
\arguments{
\item{look_up_vector}{Character vector of any bat name, or abbreviation, to
convert to scientific}

\item{bat_name_vector}{Named character vector of any bat name, or abbreviation,
against the bat species scientific name}
}
\value{
Character Vector of scientific bat names
}
\description{
A function that takes a character vector of any bat name, or abbreviation, and
compares it with a reference scientific named character vector and returns the
scientific  name of bat species using “binomial nomenclature” from the International
Code of Zoological Nomenclature (ICZN)
}
\examples{
look_up_vector <- c("45 pip" = "Pipistrellus pipistrellus", "bleb" = "Plecotus auritus")
bat_vector_to_name <- c("45 pip", "45 pip", "45 pip", "bleb")
make_scientific(look_up_vector, bat_vector_to_name)

}
