% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaryTavy.R
\docType{data}
\name{MaryTavy}
\alias{MaryTavy}
\title{Transect bat detector data as exported from Elekon's BatExplorer sound analysis software}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 244 rows and 14 columns.
}
\usage{
data(MaryTavy)
}
\description{
Results from a bat detector transect around Mary Tavy, Devon in June 2019.
data is exported 'asis' from Elekon's BatExplorer sound analysis software
}
\keyword{datasets}
