% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessyCoords.R
\docType{data}
\name{MessyCoords}
\alias{MessyCoords}
\title{A data.frame that specifies longitude and latitude coordinates in different formats.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
data(MessyCoords)
}
\description{
A data.frame with the longitude and latitude coordinates in different formats
e.g. decimal degrees and degrees minutes seconds
}
\keyword{datasets}
