% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night_sun_times.R
\name{night_sun_times}
\alias{night_sun_times}
\title{Merge Suntimes to Bat data frame}
\usage{
night_sun_times(bat_df, sun_df)
}
\arguments{
\item{bat_df}{bat data frame with Night column}
}
\value{
bat data frame with un set and sun rise times by Night
}
\description{
Merges sunset and sunrise times to the bat data frame by Night. Makes a duplicate
of sunset and shifts the the column up one row to to make the sunrise for the Night
(it will do but not yet - fudged at the moment).
}
\examples{
night_sun_times




}
