% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_of_nights.R
\name{No_Nights}
\alias{No_Nights}
\title{Number nights between bat observations}
\usage{
No_Nights(night_vector)
}
\arguments{
\item{night}{Vector of Dates}
}
\value{
Integer of the number of nights
}
\description{
Calculates the number nights between first and last bat observation (plus 1 for dawn)
}
\examples{
date_vector <- seq.Date(from=Sys.Date(), length.out = 14L, by='days')
No_Nights(date_vector)

}
