% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_data.R
\docType{data}
\name{static_data}
\alias{static_data}
\title{Static bat detector data from South Devon}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6958 rows and 5 columns.
}
\usage{
data(static_data)
}
\description{
Static bat detector data from South Devon. Co-ordinates moved slightly so as
not to disclose the location
}
\keyword{datasets}
