% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Static_G.R
\docType{data}
\name{Static_G}
\alias{Static_G}
\title{Data exported from BTO Acoustic Pipeline}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1342 rows and 19 columns.
}
\usage{
data(Static_G)
}
\description{
Anonymised data kindly provided by Richard Green Ecology
}
\keyword{datasets}
