% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statics.R
\docType{data}
\name{statics}
\alias{statics}
\title{Static bat detector data from South Devon}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6930 rows and 5 columns.
}
\usage{
data(statics)
}
\description{
Static bat detector data from a year of surveys in South Devon. Co-ordinates
moved slightly so as not to disclose the location
}
\keyword{datasets}
