% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_night_metrics.R
\name{sun_night_metrics}
\alias{sun_night_metrics}
\title{Create sun and night time metrics including bat times relative to sun set and sun rise}
\usage{
sun_night_metrics(bat_df, time_zone = "Europe/London")
}
\arguments{
\item{bat_df}{bat data frame with Night, DateTime, Latitude and Longitude columns}

\item{time_zone}{time zone location default "Europe/London"}
}
\value{
bat data frame with sun and night time metrics of bat activity relative to sun set and sun rise
}
\description{
Create night time metrics of bat activity relative to sun rise and sun set. Such as
the time of bat activity after sun set and before sun rise in minutes and hours. The activity for each hour
after sunset is also calculated along with the length of the night in hours.
}
\examples{
Night <- lubridate::ymd(c("20220620", "20220621", "20220622"))
DateTime <- lubridate::ymd_hms(c("20220620 21:30:55", "20220621 22:10:34", "20220622 21:50:28"))
Latitude <- rep(51.5023, times = 3)
Longitude <- rep(-0.1352, times = 3)
bat_data_frame <- tibble(Night, DateTime, Latitude, Longitude)
sun_night_metrics(bat_data_frame, time_zone = "Europe/London")



}
