% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_times.R
\name{sun_times}
\alias{sun_times}
\title{Sun times for the nights of bat observations}
\usage{
sun_times(bat_df, time_zone = "Europe/London")
}
\arguments{
\item{bat_df}{bat data frame with Night, Latitude and Longitude columns}
}
\value{
night and sun times data frame
}
\description{
Calculates the sunrise and sunset times for bat observations by Night (dusk to dawn). The function uses the package suncalcs to make sunset and sunrise columns for each Night and the location based on the Latitude and Longitude of the bat observation.
}
\examples{
bat_data_frame <- tibble(Night = Sys.Date(), Latitude = 51.5023, Longitude = -0.1352)
sun_times(bat_data_frame)

bat_data_frame <- tibble(Night = Sys.Date(), Latitude = -6.1370, Longitude = 106.8497)
sun_times(bat_data_frame, time_zone = "Asia/Jakarta")

}
