% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TavyOak.R
\docType{data}
\name{TavyOak}
\alias{TavyOak}
\title{Static bat detector data with call metrics from the River Tavy valley South Devon}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 626 rows and 30 columns.
}
\usage{
data(TavyOak)
}
\description{
One nights of static bat detector data, with call metrics, from the River Tavy
valley South Devon. The detector was located beneath an oak tree on a field boundary.
}
\keyword{datasets}
