% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UK_bat_names.R
\docType{data}
\name{UK_bat_names}
\alias{UK_bat_names}
\title{A table of UK bat names to aid consistent analysis}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17 rows and 5 columns.
}
\usage{
data(UK_bat_names)
}
\description{
A table of UK bat names to aid consistent analysis. The table has been adapted
from the list provided by ECOBAT http://www.ecobat.org.uk/.
}
\keyword{datasets}
