% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untidy1.R
\docType{data}
\name{untidy1}
\alias{untidy1}
\title{Untidy1 - Example of untidy bat data with too many species in a cell.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
data(untidy1)
}
\description{
Untidy1 - Example of untidy bat data with too many species in a cell.
}
\keyword{datasets}
