% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untidy2.R
\docType{data}
\name{untidy2}
\alias{untidy2}
\title{Untidy2 - Example of untidy bat data with a column giving the number of bat passes.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
data(untidy2)
}
\description{
Untidy2 - Example of untidy bat data with a column giving the number of bat passes.
}
\keyword{datasets}
