% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untidy3.R
\docType{data}
\name{untidy3}
\alias{untidy3}
\title{Untidy3 - Example of untidy bat data with two or more columns of bat species.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 4 columns.
}
\usage{
data(untidy3)
}
\description{
Untidy3 - Example of untidy bat data with two or more columns of bat species.
}
\keyword{datasets}
