% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_night_metrics.R
\name{sun_night_metrics}
\alias{sun_night_metrics}
\title{Create sun and night time metrics including bat times relative to sun set and sun rise}
\usage{
sun_night_metrics(bat_df)
}
\arguments{
\item{bat_df}{bat data frame with DateTime, sunset & sunrise columns}
}
\value{
bat data frame with sun and night time metrics of bat activity relative to sun set and sun rise
}
\description{
Create night time metrics of bat activity relative to sun rise and sun set. Such as
the time of bat activity after sun set and before sun rise in minutes and hours. The activity for each hour
after sunset is also calculated along with the length of the night in hours.
}
\examples{
sun_night_metrics




}
